requirements_cpu.txt
boto3==1.4.7
botocore==1.7.5
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
Cython==0.26
easydict==1.6
ffmpy==0.2.2
grpcio==1.8.3
grpcio-tools==1.8.3
idna==2.6
matplotlib==2.0.2
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
protobuf==3.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
scipy==0.19.1
six==1.10.0
urllib3==1.22